Documentation updates for command line parameters and e2e tests (#256)
Command line doc updates for noting that `--template-repo` can be a local path to a MLT repo and updating the `--retries` option description since it's used for both logging and `--interactive`.  Developer guide doc update to note that developers should use the `make test-e2e-no-docker` option and that `make test-e2e` option is for circleci.